feat: introduce rdfproxy.utils._typing module
The change introduces a separate rdfproxy.utils._typing
module. Type-related functions from rdfproxy.utils.mapper_utils are
moved to this module, also two new type-checking functions are added.

Given the importance of type-checking model fields in RDFProxy, the
functions of this module should be extensively unit-tested.

Concerns #223.
